Update generated queries
